feat(web): :zap: lazy load iframe
