+++ web/src/components/NewsContent.tsx
@@ -1,13 +1,18 @@
+import { useMemo } from 'react'
 import { find as linkFind } from 'linkifyjs'
 import Linkify from 'linkify-react'
 import getVideoId from 'get-video-id'
 import { Box } from '@chakra-ui/react'
 import styles from '../styles/news.module.css'
+import { VideoEmbed } from '.'
 
 const NewsContent = ({ text }: { text: string }) => {
-  const youtubeUrl = linkFind(text).find(
-    (el) => getVideoId(el.href).service === 'youtube'
-  )?.href
+  const youtubeUrl = useMemo(
+    () =>
+      linkFind(text).find((el) => getVideoId(el.href).service === 'youtube')
+        ?.href,
+    [text]
+  )
 
   return (
     <Box w="100%">
@@ -25,18 +30,8 @@ const NewsContent = ({ text }: { text: string }) => {
       </Linkify>
 
       {youtubeUrl !== undefined && (
-        <Box mt="2" w="100%">
-          <iframe
-            width="100%"
-            height="315"
-            src={`https://www.youtube-nocookie.com/embed/${
-              getVideoId(youtubeUrl).id
-            }`}
-            title="YouTube video player"
-            frameBorder="0"
-            allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
-            allowFullScreen
-          />
+        <Box mt="2">
+          <VideoEmbed url={youtubeUrl} />
         </Box>
       )}
     </Box>
