+++ web/src/components/index.ts
@@ -3,3 +3,5 @@ export { default as Navigation } from './Navigation'
 export { default as NavLink } from './NavLink'
 export { default as Footer } from './Footer'
 export { default as NewsContent } from './NewsContent'
+export { default as VideoEmbed } from './VideoEmbed'
+export { default as NewsCard } from './NewsCard'
